moved data.py to data dir
