+++ data/README.md
@@ -1,0 +1,14 @@
+<!--
+AUTHORS: RW
+-->
+
+# Usage
+View the schema of the database in `tables.py`.\
+View the functions of the corresponding class in the data controller.\
+Such as `account_data_contorller` for `accounts` class.
+
+# Function
+`write` only take class obj 
+
+TODO
+
